od: configs: Update to Buildroot 2022.11
Update the gcw0, lepus and rs90 configs to compile Binutils 2.38 instead
of 2.37, GCC 12 instead of 11, and GDB 11 instead of 10.

The python3-ssl module was also disabled, because of the hard dependency
on openssl (vs. libressl). I don't think anything uses it, so it should
be safe to remove it here.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>